refactor(entrypoint) move r2gg functions to cli file
